backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
bcrypt==4.0.1
cloudinary==1.36.0
gunicorn==21.2.0
eventlet==0.33.3
python-dotenv==1.0.0